+++ src/backend/imageGrabber/BaseX11ImageGrabber.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021 Damir Porobic <damir.porobic@gmx.com>
+ *  Copyright (C) 2016 Damir Porobic <https://github.com/damirporobic>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,20 +15,32 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor,
  * Boston, MA 02110-1301, USA.
+ *
  */
 
-#ifndef KSNIP_X11IMAGEGRABBER_H
-#define KSNIP_X11IMAGEGRABBER_H
+#ifndef X11IMAGEGRABBER_H
+#define X11IMAGEGRABBER_H
 
-#include "BaseX11ImageGrabber.h"
+#include "AbstractRectAreaImageGrabber.h"
 #include "X11Wrapper.h"
+#include "src/common/platform/HdpiScaler.h"
+#include "src/gui/snippingArea/LinuxSnippingArea.h"
 
-class X11ImageGrabber : public BaseX11ImageGrabber
+class BaseX11ImageGrabber : public AbstractRectAreaImageGrabber
 {
 public:
-	X11ImageGrabber();
-	~X11ImageGrabber() override = default;
+    explicit BaseX11ImageGrabber(X11Wrapper *x11Wrapper);
+	~BaseX11ImageGrabber() override;
+
+protected:
+	QRect fullScreenRect() const override;
+	QRect activeWindowRect() const override;
+	bool isSnippingAreaBackgroundTransparent() const override;
+	CursorDto getCursorWithPosition() const override;
+
+private:
+    X11Wrapper *mX11Wrapper;
+	HdpiScaler mHdpiScaler;
 };
 
-
-#endif //KSNIP_X11IMAGEGRABBER_H
+#endif // X11IMAGEGRABBER_H